Move urg_node to be a library.
Moving urg_node to urg_node_driver as a library.
This allows for other nodes to include this as an object instead
of spawning another separate process.
