+++ msg/Status.msg
@@ -1,4 +1,4 @@
-# Reports the status of the 'AR00' message response from the lidar.
+# Reports the status of the 'AR00' message response from the laser.
 # This contains the error state and operating modes.
 uint16 status
 
@@ -9,9 +9,9 @@ uint16 operating_mode
 
 # AKA field set
 uint16 area_number
-# If the lidar is reporting an error or not.
+# If the laser is reporting an error or not.
 bool error_status
-# THe error code the lidar is reporting.
+# The error code the laser is reporting.
 uint16 error_code
-# Does the lidar report that it is locked out.
+# Does the laser report that it is locked out.
 bool lockout_status
